Move ClientContext and ClientOptions into their own module.
